+++ src/test/java/org/opensearch/knn/index/NmslibIT.java
@@ -38,6 +38,7 @@ import java.util.TreeMap;
 import static org.hamcrest.Matchers.containsString;
 import static org.opensearch.knn.common.KNNConstants.KNN_ENGINE;
 
+@Deprecated(since = "2.19.0", forRemoval = true)
 public class NmslibIT extends KNNRestTestCase {
 
     static TestUtils.TestData testData;